refactor(www): utilize getStaticProps from nextjs-apollo-config
