+++ apps/www/pages/[...path].tsx
@@ -1,9 +1,9 @@
 import Article from '../components/Article/Page'
-import createGetStaticProps from '../lib/helpers/createGetStaticProps'
 import { GetStaticPaths } from 'next'
 import { ParsedUrlQuery } from 'querystring'
 import { gql } from '@apollo/client'
 import { getDocument } from '../components/Article/graphql/getDocument'
+import { createGetStaticProps } from '../lib/apollo/helpers'
 
 type Params = {
   path: string[]
